feat: add typescript-plugin-css-modules
