--- src/HOC/index.tsx
@@ -1,1 +1,0 @@
-export { type WithRouterProps, withRouter, routes } from 'hoc/withRouter';
