+++ src/components/shared/searchbar/components/SearchHistory.module.scss
@@ -1,0 +1,96 @@
+.histories {
+  position: absolute;
+  z-index: 2;
+  top: 50px;
+  left: 10px;
+
+  max-height: 300px;
+  width: 400px;
+  padding: 20px;
+
+  border-radius: 20px;
+  border: 2px solid var(--secondary-color);
+  outline: 2px solid var(--primary-color);
+  background: var(--primary-color);
+
+  overflow-y: auto;
+  display: inline-flex;
+  flex-wrap: wrap;
+  gap: 10px;
+
+  .history {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    column-gap: 10px;
+
+    border-radius: 20px;
+    border: 2px solid var(--primary-color);
+    outline: 2px solid var(--secondary-color);
+    background: var(--main-color);
+
+    cursor: pointer;
+    transition: outline 0.2s, box-shadow 0.2s;
+
+    &:hover {
+      box-shadow: 0 0 10px 1px var(--add-color);
+      outline: 2px solid var(--add-color);
+    }
+
+    .text {
+      padding: 2px 5px 2px 15px;
+
+      display: flex;
+      align-items: center;
+      justify-content: flex-start;
+
+      word-break: break-word;
+      line-height: 150%;
+      color: var(--primary-color);
+    }
+
+    .delete-btn {
+      position: relative;
+      right: 10px;
+      min-width: 20px;
+      min-height: 20px;
+
+      border: none;
+      border-radius: 50%;
+      background: var(--primary-color);
+
+      cursor: pointer;
+      transition: background 0.2s;
+
+      &:hover {
+        background: var(--secondary-color);
+      }
+
+      &::before {
+        content: "";
+        position: absolute;
+        top: 50%;
+        left: 50%;
+        height: 2px;
+        width: 60%;
+        transform: translate(-50%, -1px) rotateZ(45deg);
+        background: var(--main-color);
+      }
+
+      &::after {
+        content: "";
+        position: absolute;
+        top: 50%;
+        left: 50%;
+        height: 2px;
+        width: 60%;
+        transform: translate(-50%, -1px) rotateZ(135deg);
+        background: var(--main-color);
+      }
+    }
+  }
+
+  &::-webkit-scrollbar {
+    width: 0;
+  }
+}
