+++ src/components/shared/userform/components/select/SelectBasedComponent.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
 
-import styles from './SelectBasedComponent.module.scss';
-import { SelectUncontrolled } from '../CustomUncontrolledComponent';
+import { select } from './SelectBasedComponent.module.scss';
+import { SelectUncontrolled } from 'components/shared/userform/helpers/types';
 
 const SelectBasedComponent = React.forwardRef<HTMLSelectElement, SelectUncontrolled>(
   ({ id, options, placeholder }, ref) => {
     return (
-      <select ref={ref} {...{ id, name: id }} className={styles.select}>
+      <select ref={ref} {...{ id, name: id }} className={select}>
         <option value="">--- {placeholder} ---</option>
         {options?.map((option, index) => (
           <option key={`${option}_${index}`} value={option}>
