+++ src/hocs/index.tsx
@@ -1,0 +1,1 @@
+export { type WithRouterProps, withRouter, routes } from 'hocs/withRouter';
